% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_growth_pars_empty_cells.R
\name{calc_growth_pars_empty_cells}
\alias{calc_growth_pars_empty_cells}
\title{Calculate the average growth parameter within empty cells}
\usage{
calc_growth_pars_empty_cells(lu_matrices, years, cut_off_year)
}
\arguments{
\item{cut_off_year}{}
}
\description{
Calculate the average growth parameter within empty cells
}
