% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_neighbourhood_vals.R
\name{calc_neighbourhood_vals}
\alias{calc_neighbourhood_vals}
\title{Title}
\usage{
calc_neighbourhood_vals(
  index,
  lu_list,
  neigh_cols_list,
  kernel_size,
  mask,
  enr = T,
  suffix = NULL
)
}
\arguments{
\item{enr}{}
}
\description{
Title
}
