% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_lu_fact_matrices.R
\name{clean_lu_fact_matrices}
\alias{clean_lu_fact_matrices}
\title{Clean land use fractional matrices}
\usage{
clean_lu_fact_matrices(matrices_list, cut_off_year)
}
\arguments{
\item{cut_off_year}{}
}
\description{
Clean land use fractional matrices
}
