% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_data_to_country.R
\name{crop_data_to_country}
\alias{crop_data_to_country}
\title{Crop the input raster file to the extent of the country of interest}
\usage{
crop_data_to_country(
  country,
  years,
  paths_lu_filenames,
  dir_output_files,
  nc_layer = "lccs_class"
)
}
\description{
Crop the input raster file to the extent of the country of interest
}
