% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand.R
\name{demand}
\alias{demand}
\title{land use demand}
\usage{
demand(landuse, ts, path = NULL, k, type = "mean")
}
\arguments{
\item{landuse}{Matrix containing observed land use time series. Rows are cells, columns are land use classes + time steps.}

\item{ts}{Output time steps. When \code{ts} is longer than the length of the provided time series, demands for missing time steps are interpolated.}

\item{k}{Number of land use classes.}

\item{form}{Model formula (only the right hand side).}
}
\value{
A matrix containing observed and interpolated average demand for different land use classes.
}
\description{
Extracts aggregated (landscape-wide average)  land use changes from observed land use time series for model validation.
}
\examples{

}
