% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_max_dev_pars.R
\name{determine_max_dev_pars}
\alias{determine_max_dev_pars}
\title{Determines the maximum deviation parameters}
\usage{
determine_max_dev_pars(
  lu_requirements,
  lu_first_period,
  default_dev_par = 1,
  pixel_cut_off_value = 15,
  pixel_cut_off_value2 = 3,
  alt_dev_par = 5,
  alt_dev_par2 = 15
)
}
\arguments{
\item{alt_dev_par2}{}
}
\description{
Determines the maximum deviation parameters
}
