% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_by_mask.R
\name{extract_by_mask}
\alias{extract_by_mask}
\title{Align an input raster to a pre-specified reference raster}
\usage{
extract_by_mask(input_grid, ref_grid, res_method = "near")
}
\arguments{
\item{input_grid}{A SpatRaster that will be aligned}

\item{ref_grid}{A SpatRaster or sf object  that serves as a mask for the input grid. #'}

\item{res_method}{The resampling technique. Defaulting to "near", i.e.,
nearest neighbour resampling.}
}
\description{
Align an input raster to a pre-specified reference raster
}
