% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_country_names.R
\name{extract_country_names}
\alias{extract_country_names}
\title{Extract the names of the countries that may be selected as area of interest}
\usage{
extract_country_names()
}
\value{
A character with the list of country names
}
\description{
Extract the names of the countries that may be selected as area of interest
}
