% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lu_requirements.R
\name{extract_lu_requirements}
\alias{extract_lu_requirements}
\title{Extracting land use requirments}
\usage{
extract_lu_requirements(
  lu_frac_matrices_list,
  years,
  path = paste0(getwd(), "/land_use_requirements.Rds")
)
}
\arguments{
\item{path}{}
}
\description{
Extracting land use requirments
}
