% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_gridded_mask.R
\name{gen_gridded_mask}
\alias{gen_gridded_mask}
\title{Extract the indices of a gridded mask for the area of interest}
\usage{
gen_gridded_mask(
  country,
  cut_off_year,
  years,
  aggregation_factor,
  paths_lu_filenames,
  dir_output_files
)
}
\arguments{
\item{aggregation_factor}{}
}
\description{
Extract the indices of a gridded mask for the area of interest
}
