% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_panel_dataframe.R
\name{gen_panel_dataframe}
\alias{gen_panel_dataframe}
\title{Title}
\usage{
gen_panel_dataframe(
  years,
  lu_matrices,
  dir_output_files,
  country,
  aggregation_factor,
  cut_off_year = NULL
)
}
\arguments{
\item{cut_off_year}{}
}
\description{
Title
}
