% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu_allocation.R
\name{lu_allocation}
\alias{lu_allocation}
\title{Title}
\usage{
lu_allocation(
  lu,
  sm,
  params,
  dmd,
  ln,
  constraint,
  rescale,
  PA,
  output_dir = NULL
)
}
\arguments{
\item{output_dir}{}
}
\description{
Title
}
