% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhood.R
\name{neighbourhood}
\alias{neighbourhood}
\title{neihgbourhood}
\usage{
neighbourhood(lu, cols, weights, mask, ..., enr = F, suffix = NULL)
}
\arguments{
\item{lu}{land use data matrix. Each column represents a land use class, each row a cell in the landscape}

\item{cols}{Columns (land use classes) to apply neighbourhood function to}

\item{weights}{A list of weight matrices for each land use to which the neighbourhood function is applied (same length as cols)}

\item{mask}{a raster object containing the study area mask. NA values are outside of study area, non-NA values are within study area and must be of same length as the number of rows in the provided land use data matrix.}

\item{enr}{When \code{enr=TRUE}, calculated neihbourhood values are returned as enrichment factors (see ref).}
}
\value{
A matrix of neighbourhood values of the same dimensions as \code{lu}.
}
\description{
Applies neighbourhood function to data
}
\examples{

}
