% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_FLUTES.R
\name{run_FLUTES}
\alias{run_FLUTES}
\title{Run FLUTES}
\usage{
run_FLUTES(
  .country,
  .years,
  .paths_lu_filenames,
  .dir_output_files,
  .aggregation_factor,
  .path_lu_legend_filename,
  .cut_off_year = NULL,
  .lags = NULL,
  .growth_constraint = NULL,
  .max_dev = 5,
  .max_iter = 2000,
  .vals_lu_classes_to_exclude = NULL,
  .kernel_size = 25
)
}
\description{
Run FLUTES
}
