% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{allocation}
\alias{allocation}
\title{allocation}
\usage{
allocation(lu, sm, params, dmd, ln, constraint, pa = NULL)
}
\arguments{
\item{lu}{A matrix containing land use of current time step}

\item{sm}{A matrix containing the predicted land use suitability model for the current time step}

\item{params}{A list of model parameters \code{resolution}, \code{max_dev}, \code{growth}, \code{no_change} (see details).}

\item{dmd}{Matrix containing demand changes to be allocated. The first row is the current land use supply, the second row the land use demand for the next time step.}

\item{ln}{A matrix containing neighbourhood variables of the current time step.}

\item{constraint}{\code{TRUE} or \code{FALSE}. When true, constraints are applied (see details).}
}
\value{
A matrix containing predicted land use fractions.
}
\description{
Allocates changes in land use demand across the landscape
}
