% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand.R
\name{demand}
\alias{demand}
\title{land use demand}
\usage{
demand(landuse, ts, path = NULL, k, type = "mean")
}
\arguments{
\item{landuse}{Matrix containing observed land use time series. Rows correspond to cells, each column corresponds to a class and time step. This is the format in which the study's preprocessing script outputs the observed land use time series data.}

\item{ts}{Output time steps. When \code{ts} is longer than the length of the provided time series, demands for missing time steps are interpolated (linearly).}

\item{path}{Default \code{path = NULL}. Output path to write result to (.rds format).}

\item{k}{Number of land use classes.}

\item{type}{Should the landscape mean fraction be calculated (option \emph{mean}) or the number of cells changing in each class counted (option \emph{sum})?}
}
\value{
A matrix containing observed and interpolated average demand for different land use classes. Each column corresponds to a land use class, each row to a time step.
}
\description{
Extracts aggregated (landscape-wide average)  land use changes from observed land use time series for model validation.
}
