% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhood.R
\name{neighbourhood}
\alias{neighbourhood}
\title{neihgbourhood}
\usage{
neighbourhood(
  lu,
  cols,
  weights,
  mask,
  ...,
  enr = F,
  suffix = NULL,
  format = "matrix"
)
}
\arguments{
\item{lu}{land use data matrix or RasterStack object. Each column or layer represents a land use class.}

\item{cols}{Columns (land use classes) to apply neighbourhood function to}

\item{weights}{A list of weight matrices for each land use to which the neighbourhood function is applied (same length as cols)}

\item{mask}{a raster object containing the study area mask. NA values are outside of study area, non-NA values are within study area and must be of same length as the number of rows in the provided land use data matrix.}

\item{enr}{When \code{enr=TRUE}, calculated neihbourhood values are returned as enrichment factors.}

\item{format}{character string, default is \code{matrix}. When \code{format = "stack"}, neighbourhood layers are returned as raster stack.}
}
\value{
A matrix of neighbourhood values of the same dimensions and structure as \code{lu}.
}
\description{
Applies neighbourhood function to data
}
