% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_analysis.R
\name{regression_analysis}
\alias{regression_analysis}
\title{Title}
\usage{
regression_analysis(
  lu_matrices,
  neigh_values,
  lu_classes,
  lags,
  cut_off_year = NULL,
  min_n_pixels1 = 700,
  min_n_pixels2 = 400,
  min_n_pixels3 = 50,
  min_area1 = 500,
  min_area2 = 250,
  min_area3 = 10,
  cut_off_val_var = 5e-07
)
}
\arguments{
\item{lu_classes}{}
}
\description{
Title
}
