% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_output_dir.R
\name{specify_output_dir}
\alias{specify_output_dir}
\title{Specify the output directory}
\usage{
specify_output_dir(dir_output_files, aoi, aggregation_factor)
}
\arguments{
\item{aggregation_factor}{}
}
\description{
Specify the output directory
}
