% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suitmodel.R
\name{suitmodel}
\alias{suitmodel}
\title{suitability}
\usage{
suitmodel(form, lu, data, sub = NULL, resolution, ...)
}
\arguments{
\item{form}{Model formula (only the right hand side).}

\item{lu}{Matrix containing land use data}

\item{sub}{Size of subset (default \code{sub = NULL}) of data on which model is built.}

\item{resolution}{The resolution of the integer representation of proivided land use fractions (necessary in multinomial models)}

\item{...}{Further parameters passed to nnet::multinom}
}
\value{
A nnet multinomial model object
}
\description{
Estimates parameters of multinomial, multi-class suitability model
}
\examples{

}
